First draft of autotune for SuperCollider
Lots of tidying up to do, but this is a basic example of an
Antares-style autotune for SuperCollider. It uses the BACF pitch tracker
from Cycfi.

A fractional ring buffer is used to playback the input
at varying rates. As per the original effect, if the buffer is about to
overrun then the read pointer is moved back by one or more periods of
the current frequency. One would imagine this would result in horrible
clicks and discontinuities in the audio playback, however it doesn't
seem to happen very often.

Similary, if the buffer is about to underrun the read pointer is
advanced by one or more periods of the current frequency.

At present only a basic transpose param is implemented. I have some code
for tuning the input to certain keys or target pitches but it needs to
be ported over from a Bela sketch.